+++ x/tx/signing/context_test.go
@@ -145,6 +145,42 @@ func TestGetSigners(t *testing.T) {
 	}
 }
 
+func TestDefineCustomGetSigners(t *testing.T) {
+	customMsg := &testpb.Ballot{}
+	signers := [][]byte{[]byte("foo")}
+	options := Options{
+		AddressCodec:          dummyAddressCodec{},
+		ValidatorAddressCodec: dummyValidatorAddressCodec{},
+	}
+	context, err := NewContext(options)
+	require.NoError(t, err)
+
+	_, err = context.GetSigners(customMsg)
+	// without a custom signer we should get an error
+	require.ErrorContains(t, err, "use DefineCustomGetSigners to specify")
+
+	// create a new context with a custom signer
+	options.DefineCustomGetSigners(proto.MessageName(customMsg), func(msg proto.Message) ([][]byte, error) {
+		return signers, nil
+	})
+	context, err = NewContext(options)
+	require.NoError(t, err)
+	gotSigners, err := context.GetSigners(customMsg)
+	// now that a custom signer has been defined, we should get no error and the expected result
+	require.NoError(t, err)
+	require.Equal(t, signers, gotSigners)
+
+	// test that registering a custom signer for a message that already has proto annotation defined signer
+	// fails validation
+	simpleSigner := &testpb.SimpleSigner{Signer: hex.EncodeToString([]byte("foo"))}
+	options.DefineCustomGetSigners(proto.MessageName(simpleSigner), func(msg proto.Message) ([][]byte, error) {
+		return [][]byte{[]byte("qux")}, nil
+	})
+	context, err = NewContext(options)
+	require.NoError(t, err)
+	require.ErrorContains(t, context.Validate(), "a custom signer function as been defined for message SimpleSigner")
+}
+
 type dummyAddressCodec struct{}
 
 func (d dummyAddressCodec) StringToBytes(text string) ([]byte, error) {